+++ src/main/java/com/example/movix/storage/FilmStorage.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface FilmStorage {
     public Film addFilm(Film film);
-    public void removeFilm(int id);
-    public Film getById(int id);
+    public void removeFilm(Long id);
+    public Film getById(Long id);
     public List<Film> getFilms();
     public Film updateFilm(Film film);
 }
